requirements.txt
asgiref==3.7.2
Django==4.0.6
djangorestframework==3.13.1
python-dotenv==1.0.1
pytz==2024.2
sqlparse==0.4.4
typing_extensions==4.8.0
